requirements.txt
# Core dependencies
apache-airflow==2.10.4
pendulum>=3.0.0

# Database connections
pymongo>=4.0.0
psycopg2-binary>=2.9.0

# Data processing
pandas==1.5.3

# Web scraping
selenium>=4.27.0
beautifulsoup4>=4.12.0
requests>=2.32.0

# Financial data
finance-datareader>=0.9.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# Development and testing
pytest>=8.0.0
pytest-cov>=6.0.0

# Neo4j
neo4j>=6.0.2
apache-airflow-providers-neo4j>=3.10.2

# DART API
OpenDartReader>=0.2.0
